# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Social Network API",
        "description": "The public API from the Social Network team.",
        "contact": {
            "name": "the GTI525 Social Network Team."
        },
        "license": {
            "name": "MIT"
        },
        "version": "1.0.0"
    },
    "host": "https://gti525-social-network.herokuapp.com/",
    "basePath": "/api/v1",
    "schemes": [
        "http"
    ],
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
...
        "/users": {
            "get": {
                "tags": [
                    "users"
                ],
                "summary": "Show all users present in the social network database.",
                "operationId": "getUsers",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation.",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/User"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error."
                    }
                }
            }
        },
        "/user/{user_ID}": {
            "get": {
                "tags": [
                    "users"
                ],
                "summary": "Get the information about a specific user.",
                "operationId": "getUser",
                "parameters": [
                    {
                        "name": "user_ID",
                        "in": "path",
                        "description": "Target user.",
                        "required": true,
                        "type": "integer",
                        "format": "int32"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation.",
                        "schema": {
                            "$ref": "#/definitions/User"
                        }
                    },
                    "406": {
                        "description": "Not acceptable."
                    },
                    "500": {
                        "description": "Internal server error."
                    }
                }
            }
        }
    },
    "definitions": {
...
        "User": {
            "required": [
                "id",
                "first_name",
                "last_name",
                "email",
                "address",
                "city",
                "province",
                "postal_code",
                "password"
            ],
            "type": "object",
            "xml": {
                "name": "User"
            }
        }
    }
}